+++ h3_h3api.h
@@ -80,20 +80,20 @@ typedef uint32_t H3Error;
 
 typedef enum {
     E_SUCCESS = 0,  // Success (no error)
-    E_FAILED =
-        1,  // The operation failed but a more specific error is not available
-    E_DOMAIN = 2,  // Argument was outside of acceptable range (when a more
-                   // specific error code is not available)
-    E_LATLNG_DOMAIN =
-        3,  // Latitude or longitude arguments were outside of acceptable range
-    E_RES_DOMAIN = 4,    // Resolution argument was outside of acceptable range
+    E_FAILED = 1,   // The operation failed but a more specific error is not
+                    // available
+    E_DOMAIN = 2,   // Argument was outside the acceptable range (when a more
+                    // specific error code is not available)
+    E_LATLNG_DOMAIN = 3,  // Latitude or longitude arguments were outside the
+                          // acceptable range
+    E_RES_DOMAIN = 4,    // Resolution argument was outside the acceptable range
     E_CELL_INVALID = 5,  // `H3Index` cell argument was not valid
     E_DIR_EDGE_INVALID = 6,  // `H3Index` directed edge argument was not valid
     E_UNDIR_EDGE_INVALID =
         7,                 // `H3Index` undirected edge argument was not valid
     E_VERTEX_INVALID = 8,  // `H3Index` vertex argument was not valid
     E_PENTAGON = 9,  // Pentagon distortion was encountered which the algorithm
-                     // could not handle it
+                     // could not handle
     E_DUPLICATE_INPUT = 10,  // Duplicate input was encountered in the arguments
                              // and the algorithm could not handle it
     E_NOT_NEIGHBORS = 11,    // `H3Index` cell arguments were not neighbors
@@ -101,7 +101,16 @@ typedef enum {
         12,  // `H3Index` cell arguments had incompatible resolutions
     E_MEMORY_ALLOC = 13,   // Necessary memory allocation failed
     E_MEMORY_BOUNDS = 14,  // Bounds of provided memory were not large enough
-    E_OPTION_INVALID = 15  // Mode or flags argument was not valid.
+
+    E_OPTION_INVALID = 15,  // Mode or flags argument was not valid
+    E_INDEX_INVALID = 16,   // `H3Index` argument was not valid
+    E_BASE_CELL_DOMAIN =
+        17,                // Base cell number was outside of acceptable range
+    E_DIGIT_DOMAIN = 18,   // Child digits invalid
+    E_DELETED_DIGIT = 19,  // Deleted subsequence indicates invalid index
+
+    // Sentinel value; not a real error. One past the last valid code.
+    H3_ERROR_END
 } H3ErrorCodes;
 
 /** @defgroup describeH3Error describeH3Error
@@ -515,6 +524,28 @@ DECLSPEC int H3_EXPORT(getResolution)(H3Index h);
 DECLSPEC int H3_EXPORT(getBaseCellNumber)(H3Index h);
 /** @} */
 
+/** @defgroup getIndexDigit getIndexDigit
+ * Functions for getIndexDigit
+ * @{
+ */
+/** @brief returns the indexing digit of the provided H3 cell at a given
+ * resolution
+ *
+ * Indexing digits are 1-indexed beginning with the digit for resolution 1. */
+DECLSPEC H3Error H3_EXPORT(getIndexDigit)(H3Index h, int res, int *out);
+/** @} */
+
+/** @defgroup constructCell constructCell
+ * Functions for constructCell
+ * @{
+ */
+/** @brief create a cell from its components
+ * Only allows for constructing valid H3 cells.
+ **/
+DECLSPEC H3Error H3_EXPORT(constructCell)(int res, int baseCellNumber,
+                                          const int *digits, H3Index *out);
+/** @} */
+
 /** @defgroup stringToH3 stringToH3
  * Functions for stringToH3
  * @{
@@ -539,6 +570,17 @@ DECLSPEC H3Error H3_EXPORT(h3ToString)(H3Index h, char *str, size_t sz);
  * In particular, returns 0 (False) for H3 directed edges or invalid data
  */
 DECLSPEC int H3_EXPORT(isValidCell)(H3Index h);
+/** @} */
+
+/** @defgroup isValidIndex isValidIndex
+ * Functions for isValidIndex
+ * @{
+ */
+/** @brief confirms if an H3Index is valid for any mode (cell, directed edge, or
+ * vertex) Returns 1 if the H3 index is valid for any supported type, 0
+ * otherwise
+ */
+DECLSPEC int H3_EXPORT(isValidIndex)(H3Index h);
 /** @} */
 
 /** @defgroup cellToParent cellToParent